requirements.txt
appdirs 
asttokens 
attrs 
Automat 
backcall 
backports.functools-lru-cache 
bcrypt 
brotlipy 
certifi==2022.12.7
cffi 
charset-normalizer 
comm 
constantly==15.1.0
cryptography 
cssselect 
debugpy 
decorator 
dnspython==2.3.0
executing 
ipykernel
ipython 
jupyter_client 
jupyter_core 
lxml 
matplotlib-inline 
pickleshare 
platformdirs 
prompt-toolkit 
Protego 
psutil 
ptyprocess 
pure-eval 
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser 
PyDispatcher==2.0.5
Pygments 
pymongo==4.3.3
pyOpenSSL 
PySocks 
python-dateutil 
python-dotenv==1.0.0
pyzmq 
queuelib 
requests 
requests-file 
Scrapy 
service-identity==18.1.0
six 
stack-data 
tldextract 
tornado 
traitlets 
Twisted 
typing_extensions 
urllib3 
w3lib 
wcwidth 
zipp 
zope.interface 
streamlit